requirements.txt
numpy
pandas
django
pyserial
